+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 
+v0.0.4
+- Better error output
+- Maintain the indentation in the Markdown file
+
+v0.0.3
+- Minor Cargo.toml typo fix
 
 v0.0.2
-
 - Use [languages](https://github.com/cortesi/languages) crate for language detection.
+